Move DB section to other class, improve account manage validating
